requirements.txt
requests
discord.py
dotenv
audioop_lts